Make ISO SR fixtures shared
